requirements.txt
matplotlib==3.4.3
numba==0.55.1
numpy==1.18.4
Pillow==9.0.1
scipy==1.7.1
